Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/credential-management/otpcredential-iframe.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Mojo testing interface is not available.
CONSOLE MESSAGE: Unhandled Promise Rejection: Error: Mojo testing interface is not available.

FAIL Test OTPCredential enabled in same origin iframes assert_equals: expected "Pass" but got "Fail"
FAIL Test OTPCredential disabled in cross origin iframes assert_equals: expected "NotAllowedError" but got "NotSupportedError"
